Fixed some bugs and added Findbugs.

+++ src/main/scala/jenkem/engine/color/WeightedColor.scala
@@ -1,3 +1,3 @@
 package jenkem.engine.color
 
-case class WeightedColor(val ircColor: Scheme.IrcColor, val weight: Double)
+case class WeightedColor(val ircColor: Scheme.IrcColor, val weight: Float)
